requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
pyaes==1.6.1
pyextras==0.0.7
PyYAML==3.13
requests==2.21.0
tmdbsimple==2.2.0
urllib3==1.24.1
